requirements.txt
langchain-huggingface
langchain-community
langchain-text-splitters
langchain
chromadb
pypdf
pytest
boto3
python-dotenv
ollama